pyproject.toml
[project]
name = "Tesla"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "77dx", email = "396321556@qq.com"},
]
dependencies = [
    "django>=4.2.16",
    "djangorestframework>=3.15.2",
    "pytest-django>=4.9.0",
    "pygments>=2.18.0",
    "django-filter>=24.3",
    "markdown>=3.7",
    "pyyaml>=6.0.2",
    "drf-spectacular>=0.28.0",
    "ruamel-yaml>=0.18.6",
    "httpie>=3.2.4",
    "ipython>=8.12.3",
    "django-cors-headers>=4.4.0",
]
requires-python = "==3.8.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
b = 'black.'
s = 'sort.'
p = 'pytest.'

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.8.0",
    "isort>=5.13.2",
    "flake8>=5.0.4",
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "pytest-coverage>=0.0",
]
